CTF_UI/pyproject.toml
[project]
name = "ctf-ui"
version = "0.1.0"
description = "GrizzHacks8 CTF Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1",
    "flask-caching>=2.3.1",
    "flask-limiter>=4.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "frontmatter>=3.0.8",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.2",
    "argon2-cffi>=23.1.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "email-validator>=2.1.0",
    "gunicorn>=22.0.0",
    "pyperclip>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3.0",
]
